lib-shared: move metadata field name constants into PhotoMetadata
